% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMsg.R
\name{addMsg}
\alias{addMsg}
\title{addMsg}
\usage{
addMsg(
  id,
  str,
  plural = NA,
  context = NA,
  lang = getOption("gettext.lang"),
  domain = getOption("gettext.domain"),
  file = NA,
  line = NA
)
}
\arguments{
\item{id}{character: message id, see \code{msgid}}

\item{str}{character: a vector with translations, see \code{msgstr} or \code{msgstr[N]}}

\item{plural}{character: plural form of a message, see  \code{msgid_plural} (default: \code{NULL})}

\item{context}{character: context, see \code{msgctxt} (default: \code{NA})}

\item{lang}{character: language code (default: \code{getOption('gettext.lang')}, usually \code{"en"})}

\item{domain}{character: text domain (default: \code{getOption('gettext.domain')}, usually \code{NA})}

\item{file}{character: source file (default: \code{NA})}

\item{line}{integer: line number in source file (default: \code{NA})}
}
\value{
invisibly the list added to the library-internal translation environment
}
\description{
Adds a message and its translation(s) to the library-internal translation environment, see the 
\href{https://www.gnu.org/software/gettext/manual/gettext.html#PO-Files}{GNU gettext manual}.
}
\examples{
options('gettext.lang'="de") # set the current language to 'de'
addMsg("Unknown system error", "Unbekannter Systemfehler")
addMsg("Unknown system error", "Error desconegut del sistema", lang="es")
addMsg("found \%d fatal error", plural="found \%d fatal errors", 
       str=c("\%d Fehler gefunden", "\%d Fehler gefunden"))
addMsg("We saw her duck", "Wir haben ihre Ente gesehen", context="duck as noun")
addMsg("We saw her duck", "Wir haben gesehen, wie sie sich duckte", context="duck as verb")
}
