% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlural.R
\name{addPlural}
\alias{addPlural}
\title{addPlural}
\usage{
addPlural(nplurals = 1, plural = function(n) {     0 }, text = NULL,
  lang = getOption("gettext.lang"),
  domain = getOption("gettext.domain"), file = NA, line = NA)
}
\arguments{
\item{nplurals}{integer: number of plural forms (default: \code{1})}

\item{plural}{function: a function that computes the plural index from \code{n} (default: \code{function(n) {0}})}

\item{text}{character: a string from a PO file}

\item{lang}{character: language to translate to (default: \code{getOption('gettext.lang')}, usually \code{"en"})}

\item{domain}{character: text domain (default: \code{getOption('gettext.domain')}, usually \code{NA})}

\item{file}{character: source file (default: \code{NA})}

\item{line}{integer: line number in source file (default: \code{NA})}
}
\value{
invisibly the plural form stored
}
\description{
Adds a plural definition to internal translation object. It it already exists it will be overwritten.
}
\examples{
# Czech
pl <- addPlural(text='nplurals=3; plural=(n==1) ? 0 : (n>=2 && n<=4) ? 1 : 2;')
print(pl)
}
