% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind}
\alias{bind}
\title{bind}
\usage{
bind(...)
}
\arguments{
\item{...}{list: list of parameters}
}
\value{
a data frame with: \code{lang}, the assigned language, \code{domain}, the assigned domain, \code{iso639}, the identified language, and \code{file}, the filename
}
\description{
Assigns a file, based on its name or a named parameter, to a language and a domain. 
The base name of a file is composed by a \strong{domain} and \strong{language code} 
separated by non-letters. A language code starts with a two (ISO 639-1) 
letter code.
}
\details{
The result can be used by \code{readPofile}. If called with no parameters then the 
internal set of test files is given back.
}
\examples{
options('gettext.lang'='en')
bind('myproject_de.po')
bind(de='myproject.po')
bind(de=c('myproject1.po', 'myproject2.po'))
bind(de=c('myproject1_de.po', 'myproject2_de.po'), 
         en=c('myproject1_en.po', 'myproject2_en.po'))
files <- c("myproject_de.po", 
           "intl/for_use_owncloud_impress_cs_CZ.po",
           "intl/ibalsa-master-zh_HK.po",
           "spirit-po/test1.po")
bind(files)
bind()
}
