% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindtextdomain.R
\name{bindtextdomain}
\alias{bindtextdomain}
\title{bindtextdomain}
\usage{
bindtextdomain(domain = getOption("gettext.domain"), dir)
}
\arguments{
\item{domain}{character: text domain (default: \code{getOption('gettext.domain')}, usually \code{NA})}

\item{dir}{character: directory for the domain PO files}
}
\value{
a data frame with: \code{lang}, the assigned language, \code{domain}, the assigned domain, \code{iso639}, the identified language, and \code{file}, the filename
}
\description{
Sets the path for a domain.
}
\examples{
bindtextdomain('app2', system.file('shiny', 'app2', package='gettext'))
}
