% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleApp.R
\name{exampleApp}
\alias{exampleApp}
\title{exampleApp}
\usage{
exampleApp(topic = NULL)
}
\arguments{
\item{topic}{character: example(s) to show or \code{NULL}}
}
\value{
nothing
}
\description{
Runs one (or more examples) Shiny apps from the package. With \code{runExample()} you get a vector of topics back.
}
\examples{
exampleApp() 
\dontrun{
  exampleApp('app2en') 
  exampleApp('app2de') 
}
}
