% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettextify.R
\name{gettextify}
\alias{gettextify}
\title{gettextify}
\usage{
gettextify(expr, exclude = NULL)
}
\arguments{
\item{expr}{character or expression}

\item{exclude}{character: names of further function to exclude from replacement}
}
\value{
character vector with modified code
}
\description{
Takes a file, R code or R expression and replaces all text constants by \code{gettext(text_constant)}. 
The replacement is not done inside \code{gettext}, \code{G}, \code{ngettext}, \code{N}, \code{library} and
\code{require}.            
If \code{expr} is a character then \code{gettextify} tries first to open a file with the name.
If this fails then it wil be assumed that \code{expr} contains valid R code.
}
\examples{
gettextify ('x<-"a"')
gettextify (system.file(package='gettext', 'shiny', 'app1', 'app.R'))
gettextify (parse(system.file(package='gettext', 'shiny', 'app1', 'app.R')))
}
