% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTranslation.R
\name{getTranslation}
\alias{getTranslation}
\title{getTranslation}
\usage{
getTranslation(list = TRUE)
}
\arguments{
\item{list}{logical: return list or list of \code{str}'s (default: \code{TRUE})}
}
\value{
list of strings
}
\description{
Returns a copy of the internal translation environment or a list of \code{str}'s of each object.
}
\examples{
getTranslation()
}
