% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso639}
\alias{iso639}
\title{ISO 639 language codes}
\format{A data frame with 1210 language codes in three variables
\itemize{
\item \code{code} a two- or three letter code 
\item \code{english} english name of the language
\item \code{french} french name of the language
}}
\source{
\href{https://datahub.io/core/language-codes}{ISO Language Codes (639-1 and 693-2) and IETF Language Types}, 
file \href{https://datahub.io/core/language-codes/r/language-codes-full.csv}{language-codes-full.csv}
}
\usage{
data(iso639)
}
\description{
A data frame with two- and three letter codes from ISO 639-1 and ISO 693-2. 
Codes are taken from alpha-3 bibliographic, alpha-3 terminologic and alpha-2 code.
}
\examples{
data(iso639)
head(iso639)
}
\keyword{datasets}
