% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMsg.R
\name{makeKey}
\alias{makeKey}
\title{makeKey}
\usage{
makeKey(...)
}
\arguments{
\item{...}{list of strings}
}
\value{
a unique key
}
\description{
makeKey
}
\examples{
makeKey("test", 'de', '', NA)
makeKey('de', '')
}
