% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgWarning.R
\name{msgWarning}
\alias{msgWarning}
\title{msgWarning}
\usage{
msgWarning(msg, param = NULL, file = NA, line = NA)
}
\arguments{
\item{msg}{character: message}

\item{param}{chracter: character vector with additional informations or \code{NULL}}

\item{file}{character: file name or \code{NA}}

\item{line}{integer: line number or \code{NA}}
}
\value{
nothing
}
\description{
Appends a warning to the library-internal translation environment.
}
\examples{
msgWarning('duplicate entry')
}
