% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngettext.R
\name{ngettext}
\alias{ngettext}
\alias{N}
\title{ngettext}
\usage{
ngettext(id1, id2, n, lang = getOption("gettext.lang"),
  domain = getOption("gettext.domain"), context = NA)

N(id1, id2, n, lang = getOption("gettext.lang"),
  domain = getOption("gettext.domain"), context = NA)
}
\arguments{
\item{id1}{character: untranslated singular message}

\item{id2}{character: untranslated plural message}

\item{n}{integer: the number (e.g. item count) to determine the translation for the respective grammatical number}

\item{lang}{character: language code (default: \code{getOption('gettext.lang')}, usually \code{"en"})}

\item{domain}{character: text domain (default: \code{getOption('gettext.domain')}, usually \code{NA})}

\item{context}{character: context (default: \code{NA})}
}
\value{
the translated string
}
\description{
Returns a translated string from the untranslated string \code{id2}. An abbreviated form is \code{N} since 
R does not support \code{_} as unary operator.
}
\examples{
ngettext("I saw one duck", "I saw \%.0f ducks", 1)
sprintf(ngettext("I saw \%0.f duck", "I saw \%.0f ducks", 3), 3) # no translation at all
sprintf(N("I saw a duck", "I saw \%.0f ducks", 3), 3)
}
