% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plural}
\alias{plural}
\title{Plural forms}
\format{A list with 144 plural forms used in PO files 
\itemize{
\item \code{nplurals} the number of plural forms
\item \code{plural} an R function which computes from \code{n} the index of plural form
\item \code{lang} ISO 639 language code. Note that for the same language might be several plural forms possible
\item \code{form} the orignal plural form as string
}}
\source{
The plural forms are taken from the table in \href{http://docs.translatehouse.org/projects/localization-guide/en/latest/l10n/pluralforms.html}{Localization Guide 0.9.0 - Plural Forms},
}
\usage{
data(plural)
}
\description{
A list of plural forms used in PO files.
}
\examples{
data(plural)
plural[[1]]
lang <- sapply(plural, function(v) { v$lang })
# extract arabic
ar   <- which(lang=='ar')
plural[[ar]]
# test of plural form
plural[[ar]]$plural(0:100)
}
\keyword{datasets}
