% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgettext.R
\name{print.xgettext}
\alias{print.xgettext}
\title{print}
\usage{
\method{print}{xgettext}(x, ...)
}
\arguments{
\item{x}{xgettext object}

\item{...}{further parameters given to \code{print.default}}
}
\value{
returns its argument invisibly
}
\description{
print
}
\examples{
txt <- xgettext(parse(system.file('shiny', 'app1', 'app.R', package='gettext')))
txt
}
