% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printWarnings.R
\name{printWarnings}
\alias{printWarnings}
\title{printWarnings}
\usage{
printWarnings()
}
\value{
Invisibly the warnings as a character vector
}
\description{
Prints all warning messages for the translation object.
}
\examples{
printWarnings()
}
