% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTranslation.R
\name{saveTranslation}
\alias{saveTranslation}
\alias{readTranslation}
\alias{cleanTranslation}
\title{saveTranslation, readTranslation and cleanTranslation}
\usage{
saveTranslation(file, ascii = FALSE, version = 2, compress = TRUE,
  refhook = NULL)

readTranslation(file, refhook = NULL, append = TRUE)

cleanTranslation()
}
\arguments{
\item{file}{a \link{connection} or the name of the file where the \R object
    is saved to or read from.}

\item{ascii}{a logical.  If \code{TRUE} or \code{NA}, an ASCII
    representation is written; otherwise (default), a binary one is used.
    See the comments in the help for \code{\link{save}}.}

\item{version}{the workspace format version to use.  \code{NULL}
    specifies the current default version (3). The only other supported
    value is 2, the default from \R 1.4.0 to \R 3.5.0.}

\item{compress}{a logical specifying whether saving to a named file is
    to use \code{"gzip"} compression, or one of \code{"gzip"},
    \code{"bzip2"} or \code{"xz"} to indicate the type of compression to
    be used.  Ignored if \code{file} is a connection.}

\item{refhook}{a hook function for handling reference objects.}

\item{append}{logical: append to current translation environment (default: \code{TRUE})}
}
\value{
nothing
}
\description{
\itemize{
\item \code{saveTranslation} saves the library-internal translation environment as an RDS file, for details see \code{\link[base]{saveRDS}}.
\item \code{readTranslation} restores the library-internal translation environment from an RDS file, for details see \code{\link[base]{readRDS}}.
\item \code{cleanTranslation} deletes in all domains and languages the translations in the library-internal translation environment.
}
}
\examples{
\dontrun{
saveTranslation("translation.RDS")
cleanTranslation()
readTranslation("translation.RDS")
}
}
