% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePotfile.R
\name{writePotfile}
\alias{writePotfile}
\title{writePotfile}
\usage{
writePotfile(txt, file, all = TRUE, R = FALSE)
}
\arguments{
\item{txt}{xgettext object}

\item{file}{character: a file name to save the POT file}

\item{all}{logical: export all strings (default: \code{TRUE})}

\item{R}{logical: export all strings into a R or POT format (default: \code{FALSE})}
}
\value{
nothing
}
\description{
Creates a POT file from text extracted with \code{\link{xgettext}}.
}
\examples{
txt <- xgettext(parse(system.file('app1', 'app.R', package='gettext')))
\dontrun{
writePotfile(txt, 'myproject.pot')
}
}
