% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgettext.R
\name{xgettext}
\alias{xgettext}
\title{xgettext}
\usage{
xgettext(expr, ...)
}
\arguments{
\item{expr}{expression or character: expression or file name to analyse}

\item{...}{ignored}
}
\value{
a two column matrix with text constants
}
\description{
Extracts all text constants from R program including \code{gettext} calls.
}
\examples{
xgettext(system.file('shiny', 'app1', 'app.R', package='gettext'))
xgettext(system.file('shiny', 'app2', 'app.R', package='gettext'))
}
