% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPofile.R
\name{readPofile}
\alias{readPofile}
\title{readPofile}
\usage{
readPofile(file, lang = getOption("gettext.lang"),
  domain = getOption("gettext.domain"), append = FALSE)
}
\arguments{
\item{file}{character vector with file names or languageFile object}

\item{lang}{character: language code (default: \code{getOption('gettext.lang')}, usually \code{"en"})}

\item{domain}{character: text domain (default: \code{getOption('gettext.domain')}, usually \code{NA})}

\item{append}{logical: append POfile to internal translation environment or overwrite (default)}
}
\value{
invisibly the internal used languageFile object
}
\description{
Reads one or more PO file into the library internal translation object. 
If \code{file} has been created with \code{\link{language}} then several files with different
domains and languages can be read at once.
}
\examples{
\dontrun{
readPofile('test_de.po', lang='de')
readPofile(languageFiles('test_de.po'))
}

}
